Remove static resource
Create maxRevision.json when fetching scratch org, using source tracking library
