+++ packages/sfpowerscripts-cli/src/impl/pool/PoolFetchImpl.ts
@@ -7,7 +7,7 @@ import { getUserEmail } from './services/fetchers/GetUserEmail';
 import ScratchOrgInfoFetcher from './services/fetchers/ScratchOrgInfoFetcher';
 import ScratchOrgInfoAssigner from './services/updaters/ScratchOrgInfoAssigner';
 import * as fs from 'fs-extra';
-import SourceTrackingResourceController from './SourceTrackingResourceController';
+import ClientSourceTracking from './ClientSourceTracking';
 import isValidSfdxAuthUrl from './prequisitecheck/IsValidSfdxAuthUrl';
 import ScratchOrgOperator from '@dxatscale/sfpowerscripts.core/lib/scratchorg/ScratchOrgOperator';
 
@@ -131,8 +131,8 @@ export default class PoolFetchImpl extends PoolBaseImpl {
             if (isLoginSuccessFull) {
                 try {
                     const conn = (await Org.create({ aliasOrUsername: soDetail.username })).getConnection();
-                    const sourceTrackingResourceController = await SourceTrackingResourceController.create(conn, null);
-                    await sourceTrackingResourceController.retrieve();
+                    const clientSourceTracking = await ClientSourceTracking.create(conn, null);
+                    await clientSourceTracking.creatSourceTrackingFiles();
                 } catch (error) {
                     SFPLogger.log('Retriveing Source Tracking skipped.. ' + error.message, LoggerLevel.TRACE);
                 }
